refactor: move `Resetter` into `resetter` module
