+++ kernel/src/device/pci/xhci/port/mod.rs
@@ -14,6 +14,7 @@ use alloc::rc::Rc;
 use context::Context;
 use core::{cell::RefCell, convert::TryInto};
 use futures_intrusive::sync::LocalMutex;
+use resetter::Resetter;
 use x86_64::PhysAddr;
 
 mod resetter;
@@ -106,31 +107,6 @@ impl Port {
     fn read_port_rg(&self) -> PortRegisters {
         let port_rg = &self.registers.borrow().operational.port_registers;
         port_rg.read(self.index - 1)
-    }
-}
-
-struct Resetter {
-    registers: Rc<RefCell<Registers>>,
-    slot: usize,
-}
-impl Resetter {
-    fn new(registers: Rc<RefCell<Registers>>, slot: usize) -> Self {
-        Self { registers, slot }
-    }
-
-    fn reset(&mut self) {
-        self.start_resetting();
-        self.wait_until_reset_is_completed();
-    }
-
-    fn start_resetting(&mut self) {
-        let r = &mut self.registers.borrow_mut().operational.port_registers;
-        r.update(self.slot - 1, |r| r.port_sc.set_port_reset(true))
-    }
-
-    fn wait_until_reset_is_completed(&self) {
-        let r = &self.registers.borrow().operational.port_registers;
-        while !r.read(self.slot - 1).port_sc.port_reset_changed() {}
     }
 }
